Refactore APIs logic according to Pablo's advice

+++ default/methods/task/task/task_user_remove.py
@@ -8,7 +8,7 @@ from shared.database.task.task_user import TaskUser
 def api_task_user_remove(project_string_id, task_id):
     with sessionMaker.session_scope() as session:
         spec_list = [
-            {'user_id': {
+            {'user_id_list': {
             'required': False,
             }},
             {'relation': {
@@ -17,29 +17,20 @@ def api_task_user_remove(project_string_id, task_id):
                 'valid_values_list': ['reviewer', 'assignee']
             }},
         ]
-        print("remove")
-
-        result = {}
 
         log, input, untrusted_input = regular_input.master(request = request,
                                                            spec_list = spec_list)
         if len(log["error"].keys()) >= 1:
             return jsonify(log = log), 400
 
-        for user in input['user_id']:
-            users_already_assigned = session.query(TaskUser).filter(TaskUser.task_id == task_id).filter(TaskUser.relation == input['relation']).filter(TaskUser.user_id == user).count()
-            if (users_already_assigned < 1):
-                continue
-
-
-            result, log = task_user_remove_core(
-                session = session,
-                task_id = task_id,
-                user_id = user,
-                relation = input['relation'],
-                project_string_id = project_string_id,
-                log = log
-            )
+        result, log = task_user_remove_core(
+            session = session,
+            task_id = task_id,
+            user_id_list = input['user_id_list'],
+            relation = input['relation'],
+            project_string_id = project_string_id,
+            log = log
+        )
 
         if len(log["error"].keys()) >= 1:
             return jsonify(log = log), 400
@@ -49,7 +40,7 @@ def api_task_user_remove(project_string_id, task_id):
 
 def task_user_remove_core(session: 'Session',
                           task_id: int,
-                          user_id: int,
+                          user_id_list: int,
                           relation: str,
                           project_string_id: str,
                           log: dict):
@@ -60,11 +51,18 @@ def task_user_remove_core(session: 'Session',
         log['error']['project_id'] = 'Project and Task ID mismatch. Task does not belong to project.'
         return False, log
 
-    task_user = session.query(TaskUser).filter(TaskUser.task_id == task_id).filter(TaskUser.relation == relation).filter(TaskUser.user_id == user_id)
-
-    if task_user:
-        task_user.delete()
-        return True, log
-    else:
-        log['error']['task_user'] = 'Cannot find given task user relation'
+    if relation != "reviewer" and relation != "assignee":
+        log['error']['relation'] = 'Invalid relation type. Only support "reviewer", "assignee"'
         return False, log
+
+    for user_id in user_id_list:
+        users_already_assigned = session.query(TaskUser).filter(TaskUser.task_id == task_id).filter(TaskUser.relation == relation).filter(TaskUser.user_id == user_id).count()
+        if (users_already_assigned < 1):
+            continue
+
+        task_user = session.query(TaskUser).filter(TaskUser.task_id == task_id).filter(TaskUser.relation == relation).filter(TaskUser.user_id == user_id)
+
+        if task_user:
+            task_user.delete()
+
+    return True, log
